Update 001-color-overview.md
Very minor edits for punctuation and capitalization i.e. Dribble and Behance.

Suggestions are in mark tags, which *I think* will show up in preview mode. Dunno if that's too cumbersome but compared to commenting on whole lines via pull request? Aiming to make it easy to understand suggestions without needing to scour a whole line for context. Still new to this whole Github thing. So plz tell me if I'm doing it wrong :)